requirements.txt
pandas
numpy
scipy
scikit-learn==1.3.2
scikit-optimize==0.9.0
opentrons
notebook
openai
google-adk
